+++ src/thin_tts/backends/__init__.py
@@ -1,0 +1,5 @@
+"""Optional T2S inference backends."""
+
+from thin_tts.backends.manager import configure_t2s_backend
+
+__all__ = ["configure_t2s_backend"]